+++ src/main/java/io/electrum/billpay/api/AccountLookupsResource.java
@@ -17,6 +17,10 @@ import javax.ws.rs.core.UriInfo;
 import io.electrum.billpay.model.AccountLookupRequest;
 import io.electrum.billpay.model.AccountLookupResponse;
 import io.electrum.billpay.model.ErrorDetail;
+import io.electrum.billpay.model.PolicyLookupRequest;
+import io.electrum.billpay.model.PolicyLookupResponse;
+import io.electrum.billpay.model.TrafficFineLookupRequest;
+import io.electrum.billpay.model.TrafficFineLookupResponse;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -24,7 +28,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.Authorization;
 
-@Path("/accountLookups/{requestId}")
+@Path("/accountLookups")
 @Consumes({ "application/json" })
 @Produces({ "application/json" })
 @Api(description = "the accountLookups API", authorizations = { @Authorization("httpBasic") })
@@ -33,9 +37,10 @@ public abstract class AccountLookupsResource {
    protected abstract IAccountLookupsResource getResourceImplementation();
 
    @POST
+   @Path("/{requestId}")
    @Consumes({ "application/json" })
    @Produces({ "application/json" })
-   @ApiOperation(value = "BasicRequest account info", notes = "Requests the current status of the account identified by the supplied account reference. Not all services support account lookups. In the case where this function is not supported for the requested service, a 501 HTTP status code may be returned")
+   @ApiOperation(value = "Basic account information request", notes = "Requests the current status of the account identified by the supplied account reference. Not all services support account lookups. In the case where this function is not supported for the requested service, a 501 HTTP status code may be returned")
    @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = AccountLookupResponse.class),
          @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
          @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
@@ -53,6 +58,70 @@ public abstract class AccountLookupsResource {
          @Context UriInfo uriInfo) {
 
       getResourceImplementation().requestAccountInfo(
+            requestId,
+            body,
+            securityContext,
+            asyncResponse,
+            request,
+            httpServletRequest,
+            httpHeaders,
+            uriInfo);
+   }
+
+   @POST
+   @Path("/traffic/{requestId}")
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = "Basic traffic fine info request", notes = "Requests information related to a specific traffic fine as identified by the noticeNumber. In the case where this function is not supported by the relevant service provider, a 501 HTTP status code may be returned")
+   @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = TrafficFineLookupResponse.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public void requestTrafficFineInfo(
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("requestId") String requestId,
+         @ApiParam(value = "A traffic fine lookup request", required = true) TrafficFineLookupRequest body,
+         @Context SecurityContext securityContext,
+         @Suspended AsyncResponse asyncResponse,
+         @Context Request request,
+         @Context HttpServletRequest httpServletRequest,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo) {
+
+      getResourceImplementation().requestTrafficFineInfo(
+            requestId,
+            body,
+            securityContext,
+            asyncResponse,
+            request,
+            httpServletRequest,
+            httpHeaders,
+            uriInfo);
+   }
+
+   @POST
+   @Path("/policy/{requestId}")
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = "Basic policy info request", notes = "Requests the current status of the policy identified by the supplied policy reference. Not all service providers support policy lookups. In the case where this function is not supported by the requested service provider, a 501 HTTP status code may be returned")
+   @ApiResponses(value = { @ApiResponse(code = 200, message = "Success", response = PolicyLookupResponse.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public void requestPolicyInfo(
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("requestId") String requestId,
+         @ApiParam(value = "A policy lookup request", required = true) PolicyLookupRequest body,
+         @Context SecurityContext securityContext,
+         @Suspended AsyncResponse asyncResponse,
+         @Context Request request,
+         @Context HttpServletRequest httpServletRequest,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo) {
+
+      getResourceImplementation().requestPolicyInfo(
             requestId,
             body,
             securityContext,