+++ src/main/java/io/electrum/billpay/api/IPaymentsResource.java
@@ -8,6 +8,8 @@ import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 
 import io.electrum.billpay.model.PaymentRequest;
+import io.electrum.billpay.model.PolicyPaymentRequest;
+import io.electrum.billpay.model.TrafficFinePaymentRequest;
 import io.electrum.vas.model.BasicReversal;
 import io.electrum.vas.model.TenderAdvice;
 
@@ -27,6 +29,26 @@ public interface IPaymentsResource {
    public void createPayment(
          String id,
          PaymentRequest body,
+         SecurityContext securityContext,
+         AsyncResponse asyncResponse,
+         Request request,
+         HttpServletRequest httpServletRequest,
+         HttpHeaders httpHeaders,
+         UriInfo uriInfo);
+
+   public void createPayment(
+         String id,
+         TrafficFinePaymentRequest body,
+         SecurityContext securityContext,
+         AsyncResponse asyncResponse,
+         Request request,
+         HttpServletRequest httpServletRequest,
+         HttpHeaders httpHeaders,
+         UriInfo uriInfo);
+
+   public void createPayment(
+         String id,
+         PolicyPaymentRequest body,
          SecurityContext securityContext,
          AsyncResponse asyncResponse,
          Request request,