+++ src/main/java/io/electrum/billpay/api/PaymentsResource.java
@@ -17,6 +17,10 @@ import javax.ws.rs.core.UriInfo;
 import io.electrum.billpay.model.ErrorDetail;
 import io.electrum.billpay.model.PaymentRequest;
 import io.electrum.billpay.model.PaymentResponse;
+import io.electrum.billpay.model.PolicyPaymentRequest;
+import io.electrum.billpay.model.PolicyPaymentResponse;
+import io.electrum.billpay.model.TrafficFinePaymentRequest;
+import io.electrum.billpay.model.TrafficFinePaymentResponse;
 import io.electrum.vas.model.BasicAdviceResponse;
 import io.electrum.vas.model.BasicReversal;
 import io.electrum.vas.model.TenderAdvice;
@@ -28,7 +32,7 @@ import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.Authorization;
 import io.swagger.annotations.ResponseHeader;
 
-@Path("/payments/{paymentId}")
+@Path("/payments")
 @Consumes({ "application/json" })
 @Produces({ "application/json" })
 @Api(description = "the payments API", authorizations = { @Authorization("httpBasic") })
@@ -37,7 +41,7 @@ public abstract class PaymentsResource {
    protected abstract IPaymentsResource getResourceImplementation();
 
    @POST
-   @Path("/confirmations/{adviceId}")
+   @Path("/{paymentId}/confirmations/{adviceId}")
    @Consumes({ "application/json" })
    @Produces({ "application/json" })
    @ApiOperation(value = "Confirm an existing bill payment", notes = "If a createPayment request previously succeeded with a 201 status it must be confirmed or cancelled to complete the transaction. confirmPayment can only suceeed if a payment was created but not cancelled. confirmPayment must be repeated until a final HTTP status code is received (not 500 or 504). If a status code of either 500 or 504 is received, or no response is received, the request must be repeated. confirmPayment may be called repeatedly on the same payment resource without negative effect.")
@@ -70,6 +74,7 @@ public abstract class PaymentsResource {
    }
 
    @POST
+   @Path("/{paymentId}")
    @Consumes({ "application/json" })
    @Produces({ "application/json" })
    @ApiOperation(value = "Initiate a bill payment transaction", notes = "Requests that a payment be made towards a customer account")
@@ -102,7 +107,73 @@ public abstract class PaymentsResource {
    }
 
    @POST
-   @Path("/reversals/{adviceId}")
+   @Path("/traffic/{paymentId}")
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = "Initiate a traffic fine payment transaction", notes = "Requests that a payment be made towards a traffic fine.")
+   @ApiResponses(value = {
+         @ApiResponse(code = 201, message = "Created", response = TrafficFinePaymentResponse.class, responseHeaders = {
+               @ResponseHeader(name = "Location", description = "The location of the created payments resource", response = String.class) }),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public void createPayment(
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("paymentId") String paymentId,
+         @ApiParam(value = "A traffic fine payment request", required = true) TrafficFinePaymentRequest body,
+         @Context SecurityContext securityContext,
+         @Suspended AsyncResponse asyncResponse,
+         @Context Request request,
+         @Context HttpServletRequest httpServletRequest,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo) {
+
+      getResourceImplementation().createPayment(
+            paymentId,
+            body,
+            securityContext,
+            asyncResponse,
+            request,
+            httpServletRequest,
+            httpHeaders,
+            uriInfo);
+   }
+
+   @POST
+   @Path("/policy/{paymentId}")
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(value = "Initiate a policy payment transaction", notes = "Requests that a payment be made towards a policy.")
+   @ApiResponses(value = {
+         @ApiResponse(code = 201, message = "Created", response = PolicyPaymentResponse.class, responseHeaders = {
+               @ResponseHeader(name = "Location", description = "The location of the created payments resource", response = String.class) }),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public void createPayment(
+         @ApiParam(value = "The randomly generated UUID of this request", required = true) @PathParam("paymentId") String paymentId,
+         @ApiParam(value = "A policy payment request", required = true) PolicyPaymentRequest body,
+         @Context SecurityContext securityContext,
+         @Suspended AsyncResponse asyncResponse,
+         @Context Request request,
+         @Context HttpServletRequest httpServletRequest,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo) {
+
+      getResourceImplementation().createPayment(
+            paymentId,
+            body,
+            securityContext,
+            asyncResponse,
+            request,
+            httpServletRequest,
+            httpHeaders,
+            uriInfo);
+   }
+
+   @POST
+   @Path("/{paymentId}/reversals/{adviceId}")
    @Consumes({ "application/json" })
    @Produces({ "application/json" })
    @ApiOperation(value = "Reverse a payment request that failed or timed out", notes = "If a createPayment request fails with a 500 or 504 HTTP status code, or no response was received within the timeout period, it must be reversed to ensure the payment is not reflected on a customer's account. reversePayment must be repeated until a final HTTP status code is received (not 500 or 504). reversePayment may be called repeatedly on the same payment resource without negative effect.")