+++ src/main/java/io/electrum/billpay/model/BillpayResponse.java
@@ -1,6 +1,7 @@
 package io.electrum.billpay.model;
 
 import io.electrum.vas.model.Amounts;
+import io.electrum.vas.model.SlipData;
 import io.electrum.vas.model.Transaction;
 
 import javax.validation.constraints.NotNull;
@@ -10,48 +11,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 
 public abstract class BillpayResponse extends Transaction {
-   protected Account account = null;
    protected Customer customer = null;
    protected Amounts amounts = null;
    protected BillSlipData slipData = null;
-
-   /**
-    * Data that should be printed on the customer receipt
-    **/
-   public BillpayResponse slipData(BillSlipData slipData) {
-      this.slipData = slipData;
-      return this;
-   }
-
-   @ApiModelProperty(required = true, value = "Data that should be printed on the customer receipt")
-   @JsonProperty("slipData")
-   @NotNull
-   public BillSlipData getSlipData() {
-      return slipData;
-   }
-
-   public void setSlipData(BillSlipData slipData) {
-      this.slipData = slipData;
-   }
-
-   /**
-    * The customer account detail
-    **/
-   public BillpayResponse account(Account account) {
-      this.account = account;
-      return this;
-   }
-
-   @ApiModelProperty(required = true, value = "The customer account detail")
-   @JsonProperty("account")
-   @NotNull
-   public Account getAccount() {
-      return account;
-   }
-
-   public void setAccount(Account account) {
-      this.account = account;
-   }
+   protected boolean partPaymentAllowed = true;
+   protected boolean overPaymentAllowed = true;
 
    /**
     * Customer detail
@@ -88,5 +52,68 @@ public abstract class BillpayResponse extends Transaction {
 
    public void setAmounts(Amounts amounts) {
       this.amounts = amounts;
+   }
+
+   /**
+    * Data that should be printed on the customer receipt
+    **/
+   @Override
+   public BillpayResponse slipData(SlipData slipData) {
+      return slipData((BillSlipData) slipData);
+   }
+
+   /**
+    * Data that should be printed on the customer receipt
+    **/
+   public BillpayResponse slipData(BillSlipData slipData) {
+      this.slipData = slipData;
+      return this;
+   }
+
+   @ApiModelProperty(required = true, value = "Data that should be printed on the customer receipt")
+   @JsonProperty("slipData")
+   @NotNull
+   public BillSlipData getSlipData() {
+      return slipData;
+   }
+
+   public void setSlipData(BillSlipData slipData) {
+      this.slipData = slipData;
+   }
+
+   /**
+    * Indicates whether a payment amount may be less than the amount due. Defaults to true.
+    **/
+   public BillpayResponse partPaymentAllowed (boolean partPaymentAllowed) {
+      this.partPaymentAllowed = partPaymentAllowed;
+      return this;
+   }
+
+   @ApiModelProperty(value = "Indicates whether a payment amount may be less than the amount due. Defaults to true.")
+   @JsonProperty("partPaymentAllowed")
+   public boolean getPartPaymentAllowed() {
+      return partPaymentAllowed;
+   }
+
+   public void setPartPaymentAllowed(boolean partPaymentAllowed) {
+      this.partPaymentAllowed = partPaymentAllowed;
+   }
+
+   /**
+    * Indicates whether a payment amount may be more than the amount due. Defaults to true.
+    **/
+   public BillpayResponse overPaymentAllowed (boolean overPaymentAllowed) {
+      this.overPaymentAllowed = overPaymentAllowed;
+      return this;
+   }
+
+   @ApiModelProperty(value = "Indicates whether a payment amount may be more than the amount due. Defaults to true.")
+   @JsonProperty("overPaymentAllowed")
+   public boolean getOverPaymentAllowed() {
+      return overPaymentAllowed;
+   }
+
+   public void setOverPaymentAllowed(boolean overPaymentAllowed) {
+      this.overPaymentAllowed = overPaymentAllowed;
    }
 }