+++ src/main/java/io/electrum/billpay/model/PolicyLookupRequest.java
@@ -7,72 +7,50 @@ import org.hibernate.validator.constraints.Length;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.electrum.vas.Utils;
-import io.electrum.vas.model.Amounts;
 import io.electrum.vas.model.Transaction;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
 /**
- * Represents a request to perform a payment
+ * The data required to request policy info
  **/
-@ApiModel(description = "Represents a request to perform a payment")
-public class PaymentRequest extends Transaction {
+@ApiModel(description = "The data required to request policy info")
+public class PolicyLookupRequest extends Transaction {
 
-   private String accountRef = null;
-   private Amounts amounts = null;
+   private String policyNumber = null;
 
    /**
-    * A reference number identifying the bill payments processor, bill issuer, and customer
+    * A reference number identifying the policy to the service provider.
     **/
-   public PaymentRequest accountRef(String accountRef) {
-      this.accountRef = accountRef;
+   public PolicyLookupRequest policyNumber(String policyNumber) {
+      this.policyNumber = policyNumber;
       return this;
    }
 
-   @ApiModelProperty(required = true, value = "A reference number identifying the bill payments processor, bill issuer, and customer")
-   @JsonProperty("accountRef")
+   @ApiModelProperty(required = true, value = "A reference number identifying the policy to the service provider.")
+   @JsonProperty("policyNumber")
    @NotNull
    @Length(min = 6, max = 40)
-   public String getAccountRef() {
-      return accountRef;
+   public String getPolicyNumber() {
+      return policyNumber;
    }
 
-   public void setAccountRef(String accountRef) {
-      this.accountRef = accountRef;
-   }
-
-   /**
-    * Contains the payment amount.
-    **/
-   public PaymentRequest amounts(Amounts amounts) {
-      this.amounts = amounts;
-      return this;
-   }
-
-   @ApiModelProperty(required = true, value = "Contains the payment amount.")
-   @JsonProperty("amounts")
-   @NotNull
-   public Amounts getAmounts() {
-      return amounts;
-   }
-
-   public void setAmounts(Amounts amounts) {
-      this.amounts = amounts;
+   public void setPolicyNumber(String policyNumber) {
+      this.policyNumber = policyNumber;
    }
 
    @Override
    public String toString() {
       StringBuilder sb = new StringBuilder();
-      sb.append("class PaymentRequest {\n");
+      sb.append("class PolicyLookupRequest {\n");
 
       sb.append("    id: ").append(Utils.toIndentedString(id)).append("\n");
+      sb.append("    policyNumber: ").append(Utils.toIndentedString(policyNumber)).append("\n");
       sb.append("    time: ").append(Utils.toIndentedString(time)).append("\n");
       sb.append("    originator: ").append(Utils.toIndentedString(originator)).append("\n");
       sb.append("    client: ").append(Utils.toIndentedString(client)).append("\n");
       sb.append("    settlementEntity: ").append(Utils.toIndentedString(settlementEntity)).append("\n");
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
-      sb.append("    accountRef: ").append(Utils.toIndentedString(accountRef)).append("\n");
-      sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
       sb.append("    thirdPartyIdentifiers: ").append(Utils.toIndentedString(thirdPartyIdentifiers)).append("\n");
       sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    basketRef: ").append(Utils.toIndentedString(basketRef)).append("\n");