+++ src/main/java/io/electrum/billpay/model/RefundRequest.java
@@ -1,9 +1,5 @@
 package io.electrum.billpay.model;
 
-import io.electrum.vas.Utils;
-import io.electrum.vas.model.ThirdPartyIdentifier;
-import io.electrum.vas.model.Transaction;
-
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
@@ -11,6 +7,8 @@ import org.hibernate.validator.constraints.Length;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import io.electrum.vas.Utils;
+import io.electrum.vas.model.Transaction;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
@@ -78,11 +76,9 @@ public class RefundRequest extends Transaction {
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
       sb.append("    issuerReference: ").append(Utils.toIndentedString(issuerReference)).append("\n");
       sb.append("    refundReason: ").append(Utils.toIndentedString(refundReason)).append("\n");
-      sb.append("    transactionIdentifiers: [");
-      for (ThirdPartyIdentifier thirdPartyIdentifier : thirdPartyIdentifiers) {
-         sb.append(Utils.toIndentedString("\n    ")).append(Utils.toIndentedString(thirdPartyIdentifier));
-      }
-      sb.append("    ]\n");
+      sb.append("    thirdPartyIdentifiers: ").append(Utils.toIndentedString(thirdPartyIdentifiers)).append("\n");
+      sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
+      sb.append("    basketRef: ").append(Utils.toIndentedString(basketRef)).append("\n");
       sb.append("}");
       return sb.toString();
    }