+++ src/main/java/io/electrum/billpay/model/RefundResponse.java
@@ -1,14 +1,73 @@
 package io.electrum.billpay.model;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import io.electrum.vas.Utils;
-import io.electrum.vas.model.ThirdPartyIdentifier;
 import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 /**
  * Represents a response to a payment request
  **/
 @ApiModel(description = "Represents a response to a refund request")
 public class RefundResponse extends BillpayResponse {
+   protected Account account = null;
+   protected TrafficFine trafficFine = null;
+   protected Policy policy = null;
+
+   /**
+    * The customer account detail
+    **/
+   public BillpayResponse account(Account account) {
+      this.account = account;
+      return this;
+   }
+
+   @ApiModelProperty(value = "The customer account detail")
+   @JsonProperty("account")
+   public Account getAccount() {
+      return account;
+   }
+
+   public void setAccount(Account account) {
+      this.account = account;
+   }
+
+   /**
+    * The trafficFine detail
+    **/
+   public BillpayResponse trafficFine(TrafficFine trafficFine) {
+      this.trafficFine = trafficFine;
+      return this;
+   }
+
+   @ApiModelProperty(value = "The customer trafficFine detail")
+   @JsonProperty("trafficFine")
+   public TrafficFine getTrafficFine() {
+      return trafficFine;
+   }
+
+   public void setTrafficFine(TrafficFine trafficFine) {
+      this.trafficFine = trafficFine;
+   }
+
+   /**
+    * The policy detail
+    **/
+   public BillpayResponse policy(Policy policy) {
+      this.policy = policy;
+      return this;
+   }
+
+   @ApiModelProperty(value = "The customer policy detail")
+   @JsonProperty("policy")
+   public Policy getPolicy() {
+      return policy;
+   }
+
+   public void setPolicy(Policy policy) {
+      this.policy = policy;
+   }
 
    @Override
    public String toString() {
@@ -23,11 +82,15 @@ public class RefundResponse extends BillpayResponse {
       sb.append("    receiver: ").append(Utils.toIndentedString(receiver)).append("\n");
       sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
       sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
-      sb.append("    transactionIdentifiers: [");
-      for (ThirdPartyIdentifier thirdPartyIdentifier : thirdPartyIdentifiers) {
-         sb.append(Utils.toIndentedString("\n    ")).append(Utils.toIndentedString(thirdPartyIdentifier));
-      }
-      sb.append("    ]\n");
+      sb.append("    customer: ").append(Utils.toIndentedString(customer)).append("\n");
+      sb.append("    account: ").append(Utils.toIndentedString(account)).append("\n");
+      sb.append("    trafficFine: ").append(Utils.toIndentedString(trafficFine)).append("\n");
+      sb.append("    policy: ").append(Utils.toIndentedString(policy)).append("\n");
+      sb.append("    thirdPartyIdentifiers: ").append(Utils.toIndentedString(thirdPartyIdentifiers)).append("\n");
+      sb.append("    slipData: ").append(Utils.toIndentedString(slipData)).append("\n");
+      sb.append("    basketRef: ").append(Utils.toIndentedString(basketRef)).append("\n");
+      sb.append("    overPaymentAllowed: ").append(Utils.toIndentedString(overPaymentAllowed)).append("\n");
+      sb.append("    partPaymentAllowed: ").append(Utils.toIndentedString(partPaymentAllowed)).append("\n");
       sb.append("}");
       return sb.toString();
    }